Delete and Export menus

+++ source/delete.cpp
@@ -4,7 +4,26 @@
 #include "common/util.hpp"
 
 void deleteBox(Streetpass::StreetpassManager& sm, u8 slotNum) {
-    sm.DeleteBox(slotNum);
+    printf("This will delete the box in slot [%x].\n", slotNum);
+    printf("Are you sure?\n");
+    printf("A: Yes\t B: No\n");
+    hidScanInput();
+    while (aptMainLoop() && !(hidKeysDown() & KEY_A) && !(hidKeysDown() & KEY_B)) {
+        hidScanInput();
+        if (hidKeysDown() & KEY_A)
+        {
+            sm.DeleteBox(slotNum);
+            printf("Box deleted.\n");
+            waitForInput();
+            break;
+        }
+        if (hidKeysDown() & KEY_B)
+        {
+            printf("Canceled.\n");
+            waitForInput();
+            break;
+        }
+    }
 }
 
 void deleteAllBoxes(Streetpass::StreetpassManager& sm) {
@@ -18,13 +37,14 @@ void deleteAllBoxes(Streetpass::StreetpassManager& sm) {
         {
             sm.DeleteAllBoxes();
             printf("All boxes deleted.\n");
+            waitForInput();
             break;
         }
         if (hidKeysDown() & KEY_B)
         {
-            printf("Canceled\n");
+            printf("Canceled.\n");
+            waitForInput();
             break;
         }
     }
-    waitForInput();
 }
